requirements.txt
django
django-allauth
python-dotenv
Pillow